src/main/java/ca/ulaval/ift6002/m2/infrastructure/persistence/assemblers/PrescriptionDTOAssembler.java
package ca.ulaval.ift6002.m2.infrastructure.persistence.assemblers;

import ca.ulaval.ift6002.m2.domain.date.DateFormatter;
import ca.ulaval.ift6002.m2.domain.prescription.Prescription;
import ca.ulaval.ift6002.m2.infrastructure.persistence.dto.PrescriptionDTO;

public class PrescriptionDTOAssembler {

    private final DateFormatter dateFormatter;

    public PrescriptionDTOAssembler(DateFormatter dateFormatter) {
        this.dateFormatter = dateFormatter;
    }

    public PrescriptionDTO toDTO(Prescription prescription) {
        String practitioner = prescription.getPractioner().toString();
        String date = dateFormatter.dateToString(prescription.getDate());
        Integer renewals = prescription.getRenewals();

        return new PrescriptionDTO(practitioner, date, renewals);
    }
}
